chore: break down main component into sub-components
